+++ src/Z38/SwissPayment/TransactionInformation/BankCreditTransfer.php
@@ -71,9 +71,9 @@ class BankCreditTransfer extends CreditTransfer
         $creditorAccount->appendChild($this->creditorIBAN->asDom($doc));
         $root->appendChild($creditorAccount);
 
-        if ($this->hasRemittanceInformation()) {
-            $root->appendChild($this->buildRemittanceInformation($doc));
-        }
+        $this->appendPurpose($doc, $root);
+
+        $this->appendRemittanceInformation($doc, $root);
 
         return $root;
     }